refactor: Split helper into multiple files
Split helper into multiple file (per request type)

There are no code changes - just moving code in different file within same package

Close #258
